# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuickLink API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/campaigns/create": {
      "post": {
        "operationId": "createCampaign",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateCampaignInput" } } } },
        "responses": { "201": { "description": "Created campaign", "content": { "application/json": { "schema": { "type": "object", "properties": { "campaign": { "$ref": "#/components/schemas/Campaign" } }, "required": ["campaign"] } } } } }
      }
    },
...
    "/api/shorten": {
      "post": {
        "operationId": "shorten",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateShortLinkInput" } } } },
        "responses": { "201": { "description": "Created short link", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ShortLink" } } } } }
      },
      "get": {
        "operationId": "listLinks",
        "parameters": [
          { "name": "search", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "tag", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "take", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Link list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListLinksResponse" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OptimizationObjective": {
        "type": "string",
        "enum": [
          "conversion_rate",
          "revenue_per_click",
          "revenue",
          "conversion_value"
        ]
      },
      "CampaignVariant": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "name": { "type": "string" }, "destinationUrl": { "type": "string" },
          "weight": { "type": "integer" }, "enabled": { "type": "boolean" }, "isControl": { "type": "boolean" },
          "clicks": { "type": "integer" }, "conversions": { "type": "integer" }, "valueCents": { "type": "string" }
        },
        "required": ["id","name","destinationUrl","weight","enabled","isControl","clicks","conversions","valueCents"]
      },
      "Campaign": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "workspaceId": { "type": "string" }, "name": { "type": "string" },
          "slug": { "type": "string" }, "status": { "type": "string" },
          "objective": { "$ref": "#/components/schemas/OptimizationObjective" }, "version": { "type": "integer" },
          "variants": { "type": "array", "items": { "$ref": "#/components/schemas/CampaignVariant" } }
        },
        "required": ["id","workspaceId","name","slug","status","objective","version","variants"]
      },
      "CreateCampaignInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" }, "slug": { "type": "string" },
          "objective": { "$ref": "#/components/schemas/OptimizationObjective" }, "currency": { "type": "string" },
          "autoOptimize": { "type": "boolean" }, "confidenceThreshold": { "type": "integer" },
          "minSampleSize": { "type": "integer" }, "minConversions": { "type": "integer" },
          "maxTrafficShiftPercent": { "type": "integer" },
          "variants": { "type": "array", "items": { "type": "object", "properties": {
            "name": { "type": "string" }, "destinationUrl": { "type": "string" }, "weight": { "type": "integer" }, "isControl": { "type": "boolean" }
          }, "required": ["name","destinationUrl","weight"] } }
        },
        "required": ["name","slug","variants"]
      },
      "ShortLink": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "originalUrl": { "type": "string" }, "shortCode": { "type": "string" },
          "shortUrl": { "type": "string" }, "managementUrl": { "type": ["string", "null"] },
          "title": { "type": ["string", "null"] }, "description": { "type": ["string", "null"] },
          "ogImage": { "type": ["string", "null"] }, "clicks": { "type": "integer" },
          "createdAt": { "type": "string" }
        },
        "required": ["id","originalUrl","shortCode","shortUrl","clicks","createdAt"]
      },
      "ShortLinkDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "originalUrl": { "type": "string" }, "shortCode": { "type": "string" },
          "title": { "type": ["string", "null"] }, "description": { "type": ["string", "null"] },
          "ogImage": { "type": ["string", "null"] }, "tags": { "type": "array", "items": { "type": "string" } },
          "clicks": { "type": "integer" }, "createdAt": { "type": "string" }, "updatedAt": { "type": "string" },
          "expiresAt": { "type": ["string", "null"] }, "expiredUrl": { "type": ["string", "null"] },
          "maxClicks": { "type": ["integer", "null"] }, "isActive": { "type": "boolean" },
          "cloaked": { "type": "boolean" }, "webhookUrl": { "type": ["string", "null"] },
          "metaPixelId": { "type": ["string", "null"] }, "googleTagId": { "type": ["string", "null"] },
          "xPixelId": { "type": ["string", "null"] }, "riskStatus": { "type": "string" },
          "healthStatus": { "type": "string" }
        },
        "required": ["id","originalUrl","shortCode"]
      },
      "CreateShortLinkInput": {
        "type": "object",
        "properties": {
          "url": { "type": "string" }, "customCode": { "type": "string" }, "title": { "type": "string" },
          "description": { "type": "string" }, "ogImage": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "password": { "type": "string" }, "expiresAt": { "type": "string" },
          "expiredUrl": { "type": "string" }, "maxClicks": { "type": "integer" },
          "workspaceId": { "type": "string" }, "cloaked": { "type": "boolean" }
        },
        "required": ["url"]
      },
...
      "ListLinksResponse": {
        "type": "object",
        "properties": {
          "links": { "type": "array", "items": { "$ref": "#/components/schemas/ShortLinkDetail" } },
          "nextCursor": { "type": ["string", "null"] }
        },
        "required": ["links"]
      },
...
    }
  }
}