# src/app/api/v1/campaigns/create/route.ts
import {NextRequest,NextResponse} from 'next/server';import {authenticateApiKey} from '@/lib/api-keys';import {getDefaultWorkspace} from '@/lib/workspaces';import {prisma} from '@/lib/prisma';import {z} from 'zod';import {getIdempotentResponse,storeIdempotentResponse} from '@/lib/idempotency';import {rateLimit} from '@/lib/rate-limit'
export const schema=z.object({name:z.string().min(1).max(160),slug:z.string().regex(/^[a-z0-9][a-z0-9-]{1,62}$/),objective:z.enum(['conversion_rate','revenue_per_click','revenue','conversion_value']).default('conversion_rate'),variants:z.array(z.object({name:z.string().min(1),destinationUrl:z.string().url(),weight:z.number().int().min(1).max(100),isControl:z.boolean().optional()})).min(2).max(20)})
export function toCreateCampaignError(e:unknown):{error:string;status:number}{if(e instanceof z.ZodError)return{error:e.issues[0]?.message||'Invalid request',status:400};if(e instanceof Error&&/weights must total 100/i.test(e.message))return{error:e.message,status:400};return{error:'Could not create campaign',status:400}}
export async function POST(r:NextRequest){const limit=await rateLimit(r,{name:'v1-campaigns-create',limit:30,windowMs:60_000});if(!limit.allowed)return NextResponse.json({error:'Rate limit exceeded'},{status:429});const auth=await authenticateApiKey(r);if(!auth)return NextResponse.json({error:'Unauthorized'},{status:401});const w=await getDefaultWorkspace(auth.userId);if(!w)return NextResponse.json({error:'Workspace required'},{status:403});try{const body=schema.parse(await r.json());const idem=await getIdempotentResponse(r,w.id,body);if(idem?.existing)return new NextResponse(idem.existing.responseJson,{status:idem.existing.responseStatus,headers:{'Content-Type':'application/json','Idempotent-Replay':'true'}});if(body.variants.reduce((n,v)=>n+v.weight,0)!==100)return NextResponse.json({error:'Weights must total 100'},{status:400});const c=await prisma.campaign.create({data:{workspaceId:w.id,name:body.name,slug:body.slug,objective:body.objective,variants:{create:body.variants.map((v,i)=>({...v,isControl:v.isControl??i===0}))},experiments:{create:{objective:body.objective}}},include:{variants:true}});await import('@/lib/routing-config').then(({publishWorkspaceRoutingConfig})=>publishWorkspaceRoutingConfig(w.id)); const out={campaign:c};if(idem)await storeIdempotentResponse({workspaceId:w.id,keyHash:idem.keyHash,requestHash:idem.requestHash,method:r.method,path:r.nextUrl.pathname,status:201,body:out});return NextResponse.json(out,{status:201})}catch(e){const safe=toCreateCampaignError(e);if(e&&!(e instanceof z.ZodError))console.error('Campaign create failed:',e);return NextResponse.json({error:safe.error},{status:safe.status})}}